cheonji/requirements.txt
streamlit==1.44.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0
openai==1.70.0
python-dotenv==1.1.0
